Update program_2.py
changed from int to float

Ethan Steele
sept. 11

+++ program_2.py
@@ -1,10 +1,10 @@
 def average_age():
     # Get User Input
-    age1 = int(input("Enter 1st person age: "))
-    age2 = int(input("Enter 2nd persons age: "))
-    age3 = int(input("Enter 3nd persons age: "))
-    age4 = int(input("Enter 4nd persons age: "))
-    age5 = int(input("Enter 5nd persons age: "))
+    age1 = float(input("Enter 1st person age: "))
+    age2 = float(input("Enter 2nd persons age: "))
+    age3 = float(input("Enter 3nd persons age: "))
+    age4 = float(input("Enter 4nd persons age: "))
+    age5 = float(input("Enter 5nd persons age: "))
     
     # Sum ages
     results = (age1 + age2 + age3 + age4 + age5)
@@ -13,4 +13,4 @@ def average_age():
     # Print the results
     print("the result is:", results_f)
 # Line which calls the above function.
-average_age()
+average_age()
